pyproject.toml
[project]
name = "repl-nix-imagetotextextractor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "pytesseract>=0.3.13",
    "streamlit>=1.40.1",
]
